+++ tests/test_access.py
@@ -5,6 +5,10 @@ def test_compound() -> None:
     compound = TagCompound("", [TagInt("foo", 1), TagInt("bar", 2)])
     assert compound["foo"] == TagInt("foo", 1)
     assert compound["bar"] == TagInt("bar", 2)
+    assert compound.get("foo") == TagInt("foo", 1)
+    assert compound.get("bar", default=TagInt("boo", 42)) == TagInt("bar", 2)
+    assert compound.get("baz", default=TagInt("boo", 42)) == TagInt("boo", 42)
+    assert compound.get("baz") is None
 
 
 def test_list() -> None: